% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensSurv.R
\name{add}
\alias{add}
\title{Ploting results of sensitivity analyisis}
\usage{
add(x, y)
}
\arguments{
\item{data}{A matched sample}

\item{exp}{A variables defining exposure group}

\item{outcome}{The outcome variable}

\item{failtime}{Time to event}

\item{Gamma}{Bias to be assessed}

\item{Gammainterval}{interval between two hidden bias to be assessed}

\item{alpha}{Significance level}
}
\description{
This function allows you to assess how sensitive your results are to unmeasured variable.
}
\references{
Section 4.4.8. of Rosenbaum PR (2002) Observational Studies, 2nd Edition.
}
\keyword{Sensitivity}

