% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensSurv.R
\name{ampPlot}
\alias{ampPlot}
\title{Amplifying and visualizing gamma parameter}
\usage{
ampPlot(gamma, lambda)
}
\arguments{
\item{gamma}{gamma parameter to be amplfied.}

\item{lambda}{vector of n-fold increase in odds of treatment.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Amplifying and visualizing gamma parameter
}
\examples{
ampPlot(10, 1)
}

