% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{contSens}
\alias{contSens}
\title{Sensitivity Test for Continous outcomes}
\usage{
contSens(x, y = NULL, data = NULL, exp = NULL, outcome = NULL,
  CausalEst = NULL, Gamma = NULL, GammaInc = NULL, alpha = NULL,
  plot_title = NULL)
}
\arguments{
\item{x}{Treatment group outcomes in same order as treatment group outcomes or an objects from a Match,matchit or bmatch package.}

\item{y}{Control group outcomes in same order as treatment group outcomes unnecessary when using a Match,matchit or bmatch packages object.}

\item{exp}{treatment variable. Must be specified for MatchIt objects not required for match objects.}

\item{outcome}{treatment variable. Must be specified for MatchIt objects not required for match objects.}

\item{Gamma}{Upper-bound on gamma parameter.}

\item{GammaInc}{To set user-specified increments for gamma parameter.}

\item{alpha}{significance level.}

\item{plot_title}{plot title.}
}
\description{
This function performs sensitivity analysis for continious outcomes in observational studies
}
\examples{
library(ggplot2);library(Matching);library(MatchIt)
data(toy)

## Creating Matching object
psmodel <- glm(treated ~ covA + covB + covC + covD + covE + covF + Asqr + BC + BD, family=binomial(), data=toy)
toy$linps = psmodel$fitted.values
X <- toy$linps
Tr <- as.logical(toy$treated)
Y = toy$out1.cost
match1 <- Match(Y=Y,Tr=Tr, X=X, M = 1, replace=FALSE, ties=FALSE)

contSens(x=match1, Gamma=5,GammaInc = 0.1,alpha = 0.05,plot_title = 'Time To Event Outcome Sensitivity Plot') #using mathit object

## Creating MatchIt object
match.it <- matchit(treated ~ covA + covB + covC + covD + covE + covF + Asqr + BC + BD, data = toy, method='nearest', ratio=1)

contSens(x=match.it ,exp = 'treated',outcome = 'out1.cost', CausalEst = 15.5,Gamma=5,GammaInc = 0.1,alpha = 0.05,plot_title = 'Time To Event Outcome Sensitivity Plot') #using mathit object


}
\author{
David Ngendahimana, Case Western Reserve University.
}
\references{
Rosenbaum, Paul R. (2002) Observational Studies. Springer-Verlag.
}

