% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{ds_function}
\alias{ds_function}
\title{Estimating design sensitivity.}
\usage{
ds_function(x, sf = 1/3, Card_Dat = NULL)
}
\arguments{
\item{x}{A list of matched sampling algorithms.}

\item{sf}{A fraction used to split the entire matched sample into planning sample (to estimate design sensitivity) and analysis sample (for outcome analysis).}

\item{Card_Dat}{Dataframe passed to the function implementing cardinality matching. Passed to \code{ds_function} only if cardinality matching is one of the matching algorithms to be evaluated}
}
\description{
This function estimates design sensitivity parameter associated with a list of matching algorithms.
}
\examples{
arg.list1 = list("a"=m.out_Nrst , "b"=m.out_NrstCAL)
arg.list2= list("a"=m.out_Nrst , "b"=m.out_NrstCAL, "c"=m.out_dsnMatch, "d"=m.out_optimal)
k=ds_function(arg.list2,Card_Dat = Card_Dat)
k$designSensitivity
}

