% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsn_data}
\alias{dsn_data}
\title{Simulated dataset to illustrate how \code{ds_function} in the SensitivityR5 package works}
\format{A data frame with 100 rows and 11 variables. 300 treated  and 700 controls :
\describe{
  \item{V1 - V6, V8}{Continious variables}
  \item{V7,V9}{Binary variables}
  \item{Y2}{Outcome}
  \item{exposure}{Treatment indicator}
  .
  .
  .
}}
\usage{
dsn_data
}
\description{
The data set has one outcome, one exposure and 9 covariates, with no missing observations anywhere.
}
\keyword{datasets}

