% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{edaTable}
\alias{edaTable}
\title{Creating table One}
\usage{
edaTable(baselinevars, outcomeVar, data)
}
\arguments{
\item{x}{Count of pairs where only control has outcome.}

\item{y}{Count of pairs where onlytreated has outcome.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Creating table One
}
\examples{
edaTable (220, 380)
}

