% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{multiControlSens}
\alias{multiControlSens}
\title{Sensitivity Analysis with multiple controls}
\usage{
multiControlSens(X, outcomeName, Gamma, GammaInc, n_contrl)
}
\arguments{
\item{X}{A `MatchIt` object}

\item{outcomeName}{Name of the outcome.}

\item{Gamma}{Upper bound of sensitivity parameter}

\item{GammaInc}{Interval width of sequence between 1 and Gamma.}

\item{n_contrl}{Number of control matched to each treated observation}
}
\description{
Sensitivity Analysis with multiple controls
}
\examples{
data("lalonde",package ="MatchIt")
m.out <- matchit(f.build("treat", covs0), data = lalonde, method = "nearest", replace = TRUE,ratio = 2)
 multiControlSens(X =m.out,outcomeName = "re78",Gamma = 2,GammaInc = 0.1,n_contrl = 2)
}

