% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{pens2}
\alias{pens2}
\title{Sensitivity analysis with Matching, MatchIt and designmatch objects.}
\usage{
pens2(x, y = NULL, est = NULL, Gamma = 2, GammaInc = 0.1, data = NULL,
  treat)
}
\arguments{
\item{x}{Treatment group outcomes or an objects from a Match,MatchIt or designmatch.}

\item{y}{Control group outcomes in same order as treatment group outcomes such that members of a pair occupy the same row in both x and y. Should not be specified x is a Matching, MatchIt and designmatch objects.}

\item{est}{Treatment effect. Must be specified if x is not a MatchIt object.}

\item{Gamma}{Upper bound of sensitivity parameter}

\item{GammaInc}{interval width for increasing gamma from 1 until the specified upper bound of sensitivity parameter is reached.}

\item{data}{Dataframe used to during matching. You do not have to specify this parameter if x is a MatchIt object}

\item{treat}{Treatmetn/Exposure variable name.}
}
\value{
a table of Rosenbaum bounds
}
\description{
Sensitivity analysis with Matching, MatchIt and designmatch objects.
}
\examples{
data("lalonde",package = "Matching")
m.out = matchit(treat ~ age  + educ +  black + hisp +married + nodegr + re74  + re75  +
u74 + u75, family=binomial, data = lalonde, method = "nearest")
mod = lm(re78~age  + educ +  black + hisp +married + nodegr + re74  + re75  +u74 + u75,data = match.data(m.out))
pens2(x = m.out, y="re78",Gamma = 2, GammaInc = 0.1,est = 629.7)
}

