% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rhc}
\alias{rhc}
\title{Right Heart Catheterization Dataset}
\format{A data frame with 5735 rows and 63 variables:
\describe{
  \item{Age}{Age, in years}
  \item{Sex}{Sex at birth}
  \item{Income}{Yearly Income, in dollars}
  \item{meta}{Metabolic Diagnostic}
  .
  .
  .
}}
\source{
\url{http://biostat.mc.vanderbilt.edu/wiki/pub/Main/DataSets/rhc.html}
}
\usage{
rhc
}
\description{
This dataset was used in Connors et al. (1996): The effectiveness of RHC in the initial care of critically
 #' ill patients. J American Medical Association 276:889-897. The dataset pertains to day 1 of hospitalizatio#' n, i.e., the 'treatment' variable swang1 is whether or not a patient received a RHC (also called the Swan
 #'-Ganz catheter) on the first day in which the patient qualified for the SUPPORT study (see above)
}
\keyword{datasets}

