% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{rubinRules2}
\alias{rubinRules2}
\title{Assessing the three Rubin Rules.}
\usage{
rubinRules2(data, Treatment, matchscore = "ps", covlist)
}
\arguments{
\item{data}{data set to be used.}

\item{Treatment}{A variables defining exposure group.}

\item{matchscore}{Variable containing matching distance.Default is propensity score.}

\item{covlist}{list of variables to be balanced. Note: All variable should be of numeric type.}
}
\description{
This function allows you to assess how sensitive your results are to unmeasured variable.
}
\examples{
data(toy)
psmodel <- glm(treated ~ covA + covB + covC + covD + covE + covF + Asqr + BC + BD, family=binomial(), data=toy)

toy$ps <- psmodel$fitted
toy$linps <- psmodel$linear.predictors

covlist1=c('covA', 'covB', 'covC', 'covD', 'covE', 'covF.Middle', 'covF.High', 'Asqr','BC', 'BD')

rubinRules(data=toy,Treatment='treated',covlist=covlist1)
}

