% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensSurv.R
\name{sensbin}
\alias{sensbin}
\title{Visualizing sensivity results}
\usage{
sensbin(data, match_f, object_name, exposure, outcome, Gamma, GammaInc,
  alpha = 0.05)
}
\arguments{
\item{data}{Dataset before matching.}

\item{match_f}{matching function that was used. Accepts match, matchit or bimatch only.}

\item{object_name}{name of the \code{match},\code{matchit} and \code{bimatch} object.}

\item{exposure}{name of the exposure variable. This name must be in quotation marks.}

\item{outcome}{name of the outcome variable. This name must be in quotation marks.}

\item{Gamma}{Sensitivity parameter.}

\item{GammaInc}{Increamental value to be used in generating Sensitivity table}

\item{alpha}{Significant value. Default is 0.05}
}
\description{
Visualizing sensivity results
}
\examples{
data(rhc)
rhc1 = rhc[,c("swang1","age", "female", "edu", "income", "ninsclas", "race", "cat1", "dnr1", "wtkilo1","hrt1", "meanbp1", "resp1", "temp1", "card", "gastr", "hema", "meta", "neuro", "ortho","renal", "resp", "seps", "trauma" ,"amihx", "ca","cardiohx" ,"chfhx", "chrpulhx","dementhx" ,"gibledhx","immunhx", "liverhx", "malighx", "psychhx","renalhx", "transhx","aps1", "das2d3pc","scoma1", "surv2md1","alb1", "bili1", "crea1", "hema1", "paco21", "pafi1", "ph1","pot1", "sod1","urin1.NA", "urin1.i", "wblc1","surv_30")]
# Creating Matchit object.
m.out1 <- matchit(swang1 ~ age + female + edu + income + ninsclas + race + cat1 + dnr1 + wtkilo1 + hrt1 + meanbp1 + resp1 + temp1 + card + gastr + hema + meta + neuro + ortho + renal + resp + seps + trauma +amihx + ca + cardiohx + chfhx + chrpulhx + dementhx + gibledhx + immunhx + liverhx + malighx + psychhx + renalhx + transhx + aps1 + das2d3pc + scoma1 + surv2md1 + alb1 + bili1 + crea1 + hema1 + paco21 +pafi1 + ph1 + pot1 + sod1 + urin1.NA + urin1.i + wblc1,data = rhc1)
sensbin (data= rhc1,match_f ="matchit",object_name=m.out1,exposure = "swang1",outcome = "surv_30",Gamma = 1.5,GammaInc = 0.01)
}

