% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensSurv.R
\name{sensSurv}
\alias{sensSurv}
\title{Sensitivity Analysis for A Simple Comparison for Censored Survival}
\usage{
sensSurv(data, exp, outcome, failtime, Gamma, alpha, Gammainterval)
}
\arguments{
\item{data}{A matched sample}

\item{exp}{A variables defining exposure group}

\item{outcome}{The outcome variable}

\item{failtime}{Time to event}

\item{Gamma}{Bias to be assessed}

\item{alpha}{Significance level}

\item{Gammainterval}{interval between two hidden bias to be assessed}
}
\description{
This function allows you to assess how sensitive your results are to unmeasured variable.
}
\references{
Section 4.4.8. of Rosenbaum PR (2002) Observational Studies, 2nd Edition.
}
\keyword{Sensitivity}

