% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensSurv.R
\name{subtract}
\alias{subtract}
\title{Visualizing sensivity results}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{Count of pairs where only control has outcome.}

\item{y}{Count of pairs where onlytreated has outcome.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Visualizing sensivity results
}
\examples{
subtract (1, 1)
subtract (10, 1)
}

