% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy}
\alias{toy}
\title{Toy dataset for EPBI500 at Case Western Reserve University}
\format{A data frame with 200 rows and 6 variables. 70 treated - subjects 131-200 and 130 controls - subjects 1-130:
\describe{
  \item{CovA}{Age, in years}
  \item{CovB}{Sex at birth}
  \item{CovC}{Yearly Income, in dollars}
  \item{CovD}{Metabolic Diagnostic}
  .
  .
  .
}}
\source{
\url{https://sites.google.com/a/case.edu/love-500/home/data-and-code}
}
\usage{
toy
}
\description{
The Data Set is 100\% fictional. It has three outcomes, with no missing observations anywhere. We  assume that a logical argument suggests that the square of `covA`, as well as the interactions of `covB` with `covC` and with `covD` should be related to treatment assignment, and thus should be  included in our propensity model.
}
\keyword{datasets}

