% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{ampPlot}
\alias{ampPlot}
\title{Amplifying and visualizing gamma parameter}
\usage{
ampPlot(gamma, lambda)
}
\arguments{
\item{gamma}{gamma parameter to be amplfied.}

\item{lambda}{vector of 3 possible n-fold increase in odds of treatment to be amplified.}
}
\value{
An amplification plot.
}
\description{
Amplifying and visualizing gamma parameter
}
\examples{
ampPlot(10, c(12,13,30))
}

