% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{rubinRules}
\alias{rubinRules}
\title{Assessing the three rubin rules}
\usage{
rubinRules(data, Treatment, matchscore = "ps", covlist)
}
\arguments{
\item{data}{A matched sample}

\item{exp}{A variables defining exposure group}

\item{outcome}{The outcome variable}

\item{failtime}{Time to event}

\item{Gamma}{Bias to be assessed}

\item{Gammainterval}{interval between two hidden bias to be assessed}

\item{alpha}{Significance level}
}
\description{
This function allows you to assess how sensitive your results are to unmeasured variable.
}
\examples{
data(toy)
psmodel <- glm(treated ~ covA + covB + covC + covD + covE + covF + Asqr + BC + BD, family=binomial(), data=toy)

toy$ps <- psmodel$fitted
toy$linps <- psmodel$linear.predictors

covlist1=c('covA', 'covB', 'covC', 'covD', 'covE', 'covF.Middle', 'covF.High', 'Asqr','BC', 'BD')

rubinRules(data=toy,Treatment='treated',covlist=covlist1)
}
\references{
Section 4.4.8. of Rosenbaum PR (2002) Observational Studies, 2nd Edition.
}
\keyword{Sensitivity}

