% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensSurv.R
\name{Survsens}
\alias{Survsens}
\title{Sensitivity Test for Matched Time to Event Outcomes}
\usage{
Survsens(data = toy, object_name = match1, exp, outcome, failtime, Gamma,
  alpha, Gammainterval, plot_title = NULL)
}
\arguments{
\item{data}{dataset.}

\item{object_name}{name of the matching object. Should be either a \code{Match} or \code{matchit} object.}

\item{exp}{name of treatment variable.}

\item{outcome}{name of outcome variable.}

\item{failtime}{name of event time variable.}

\item{Gamma}{upper bound for sensitivity parameter.}

\item{Gammainterval}{value by which to increament sensitivity parameter from zero to \code{Gamma}}

\item{plot_title}{main title of your plot.}
}
\description{
This function performs sensitivity analysis for time to event outcomes in observational studies
}
\examples{
data(toy)
psmodel <- glm(treated ~ covA + covB + covC + covD + covE + covF + Asqr + BC + BD, family=binomial(), data=toy)
toy$ps <- psmodel$fitted
toy$linps <- psmodel$linear.predictors
X <- toy$linps ## matching on the linear propensity score
Tr <- as.logical(toy$treated)
Y <- toy$out3.time
match1 <- Match(Y=Y, Tr=Tr, X=X, M = 1, replace=FALSE, ties=FALSE)
match.it <- matchit(treated ~ covA + covB + covC + covD + covE + covF + Asqr + BC + BD, data = toy, method='nearest', ratio=1)
sensSurv(data =toy,exp='treated',outcome = 'out2',failtime = 'out3.time',Gamma=2.4,Gammainterval = 0.01,alpha = 0.05,object_name = match.it,plot_title = 'Time To Event Outcome Sensitivity Plot')

}
\references{
Rosenbum 2011
}

